feat: Stage 8C-V1-1D and 8C-T2 - Semantic Vendor Boundary and Tax Observation Layer
This commit consolidates the remediation of the legacy scalar scorer boundary and the structural decoupling of tax identifier observation from vendor semantics.

Key Architectural Changes:
1. Decision Engine Coverage Enforcement (8C-V1-1D)
- Enforced minimum_coverage_threshold (0.80) inside DecisionPolicy.
- Bypassed historical risk of sparse-evidence positive inflation.
- Bridged VendorEvidenceProjection.contradiction_markers to PairScoringResult.
- Introduced ONE_TO_ONE_BLOCKING_FINDINGS for PAN/Legal Form conflicts.

2. Shared Identifier Observation Layer (8C-T2)
- Created TaxIdentifierObservation ontology (TaxIdentifierCandidateType, TaxObservationState).
- Implemented DeterministicTaxParser to isolate structural regex evaluation for GSTIN/PAN candidates.
- Created ParsedTaxIdentifierArtifact.

3. Vendor Pipeline Decoupling (8C-T2)
- VendorNameObservation and DeterministicVendorParser now delegate tax extraction to DeterministicTaxParser.
- Preserved legacy tax_identity_score (0.0/1.0 fallback) boundary in core_providers to protect engine outcomes until the next semantic phase.

Validation:
- Added comprehensive metamorphic audit documentation (docs/architecture, docs/algorithms, docs/testing).
- End-to-end 501 test suite passes seamlessly without legacy regression.

+++ experiments/compare_tax_penalty_models.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 import csv
 from dataclasses import dataclass
 from datetime import date
@@ -75,7 +76,7 @@ def parse_purchase_record(row: dict[str, str]) -> PurchaseRecord:
     return PurchaseRecord(record_id="pur_cedf16a5", 
         vendor_name=optional_text(row.get("vendor_name", "")),
         reference=optional_text(row.get("invoice_number", "")),
-        amount=float(row["amount"]),
+        amount=Decimal(row["amount"]),
         record_date=date.fromisoformat(row["invoice_date"]),
         tax_identity=optional_text(row.get("gstin", "")),
     )
@@ -85,7 +86,7 @@ def parse_gst_record(row: dict[str, str]) -> GSTRecord:
     return GSTRecord(record_id="gst_29b08642", 
         vendor_name=optional_text(row.get("supplier_name", "")),
         reference=optional_text(row.get("invoice_number", "")),
-        amount=float(row["amount"]),
+        amount=Decimal(row["amount"]),
         record_date=date.fromisoformat(row["invoice_date"]),
         tax_identity=optional_text(row.get("gstin", "")),
     )
