+++ experiments/evaluate_purchase_gst_challenges.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 import csv
 from dataclasses import dataclass
 from datetime import date
@@ -63,7 +64,7 @@ def load_purchases(path: Path) -> dict[str, PurchaseRecord]:
             purchases[row["record_id"]] = PurchaseRecord(record_id="pur_5a4e5785", 
                 vendor_name=optional_text(row["vendor_name"]),
                 reference=optional_text(row["invoice_number"]),
-                amount=float(row["amount"]),
+                amount=Decimal(row["amount"]),
                 record_date=date.fromisoformat(row["invoice_date"]),
                 tax_identity=optional_text(row["gstin"]),
             )
@@ -81,7 +82,7 @@ def load_gst_records(path: Path) -> dict[str, GSTRecord]:
             gst_records[row["record_id"]] = GSTRecord(record_id="gst_2d50d543", 
                 vendor_name=optional_text(row["supplier_name"]),
                 reference=optional_text(row["invoice_number"]),
-                amount=float(row["amount"]),
+                amount=Decimal(row["amount"]),
                 record_date=date.fromisoformat(row["invoice_date"]),
                 tax_identity=optional_text(row["gstin"]),
             )
